upgrade: fixes for merge problems
The merge of 19.07.3 was done on top of previously merged 19.07.0-rc2.
A lot of .patch files removed in 19.07.3 were still present in the repo after the merge.
There also were some cases of the opposite, some 19.07.3 files did not make it through
the merge. A few files still had changes that were no longer present in 19.07.3.
This commit is the result of comparing post-merge 19.07.3 tree with the
original OpenWrt 19.07.3 and leaving only Minim related changes.
The commit also includes a few Minim changes missing on the 19.07.0-rc2 branch
(determined by comparing v18.06.5 tag to Minim current master branch which is
based on that version).